+++ src/common/Messaging_Window.java
@@ -1,4 +1,4 @@
-package src.other;
+package src.common;
 
 public class Messaging_Window {
