+++ src/resources/fxml/Connection_Settings.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 <?import java.net.URL?>
-<StackPane prefHeight="158.0" prefWidth="366.0" xmlns="http://javafx.com/javafx/8.0.111" xmlns:fx="http://javafx.com/fxml/1" fx:controller="src.other.Connection_Settings_Controller">
+<StackPane prefHeight="158.0" prefWidth="366.0" xmlns="http://javafx.com/javafx/8.0.111" xmlns:fx="http://javafx.com/fxml/1" fx:controller="src.common.Connection_Settings_Controller">
    <children>
       <VBox prefHeight="200.0" prefWidth="600.0">
          <children>
